modify: move P2PMessageRC2->P2PMessage, P2PMessage->P2PMessageRC1 && make P2PMessageRC1 override from P2PMessage
