+++ libp2p/P2PMessage.cpp
@@ -1,4 +1,3 @@
-
 /*
  * @CopyRight:
  * FISCO-BCOS is free software: you can redistribute it and/or modify
@@ -22,25 +21,71 @@
 
 #include "P2PMessage.h"
 #include "Common.h"
+#include <libconfig/GlobalConfigure.h>
+#include <libdevcore/SnappyCompress.h>
 
 using namespace dev;
 using namespace dev::p2p;
+using namespace dev::compress;
 
 void P2PMessage::encode(bytes& buffer)
 {
+    std::shared_ptr<bytes> compressData = std::make_shared<bytes>();
+    /// compress success
+    if (compress(compressData))
+    {
+        encode(buffer, compressData);
+    }
+    else
+    {
+        encode(buffer, m_buffer);
+    }
+}
+
+/**
+ * @brief: encode the message into bytes to send
+ * @param buffer: the encoded data
+ * @param encodeBuffer: the message should be encoded
+ * @param protocol: the protocol ID that should be encoded into the buffer
+ */
+void P2PMessage::encode(bytes& buffer, std::shared_ptr<bytes> encodeBuffer)
+{
     buffer.clear();  ///< It is not allowed to be assembled outside.
-    m_length = HEADER_LENGTH + m_buffer->size();
+    m_length = HEADER_LENGTH + encodeBuffer->size();
 
     uint32_t length = htonl(m_length);
-    PROTOCOL_ID protocolID = htons(m_protocolID);
+    VERSION_TYPE versionType = htons(m_version);
+    PROTOCOL_ID protocolID = htonl(m_protocolID);
     PACKET_TYPE packetType = htons(m_packetType);
     uint32_t seq = htonl(m_seq);
 
-    buffer.insert(buffer.end(), (byte*)&length, (byte*)&length + 4);
-    buffer.insert(buffer.end(), (byte*)&protocolID, (byte*)&protocolID + 2);
-    buffer.insert(buffer.end(), (byte*)&packetType, (byte*)&packetType + 2);
-    buffer.insert(buffer.end(), (byte*)&seq, (byte*)&seq + 4);
-    buffer.insert(buffer.end(), m_buffer->begin(), m_buffer->end());
+    buffer.insert(buffer.end(), (byte*)&length, (byte*)&length + sizeof(length));
+    buffer.insert(buffer.end(), (byte*)&versionType, (byte*)&versionType + sizeof(versionType));
+    buffer.insert(buffer.end(), (byte*)&protocolID, (byte*)&protocolID + sizeof(protocolID));
+    buffer.insert(buffer.end(), (byte*)&packetType, (byte*)&packetType + sizeof(packetType));
+    buffer.insert(buffer.end(), (byte*)&seq, (byte*)&seq + sizeof(seq));
+    buffer.insert(buffer.end(), encodeBuffer->begin(), encodeBuffer->end());
+}
+
+/// compress the data to be sended
+bool P2PMessage::compress(std::shared_ptr<bytes> compressData)
+{
+    if (!g_BCOSConfig.compressEnabled() || m_buffer->size() <= g_BCOSConfig.c_compressThreshold)
+    {
+        return false;
+    }
+    /// the packet has already been encoded
+    if ((m_version & dev::eth::CompressFlag) == dev::eth::CompressFlag)
+    {
+        return false;
+    }
+    size_t compressSize = SnappyCompress::compress(ref(*m_buffer), *compressData);
+    if (compressSize < 1)
+    {
+        return false;
+    }
+    m_version |= dev::eth::CompressFlag;
+    return true;
 }
 
 ssize_t P2PMessage::decode(const byte* buffer, size_t size)
@@ -53,59 +98,36 @@ ssize_t P2PMessage::decode(const byte* buffer, size_t size)
     int32_t offset = 0;
     m_length = ntohl(*((uint32_t*)&buffer[offset]));
 
-    /*if (m_length > MAX_LENGTH)
-    {
-        return PACKET_ERROR;
-    }*/
-
     if (size < m_length)
     {
         return dev::network::PACKET_INCOMPLETE;
     }
-
-    offset += 4;
-    m_protocolID = ntohs(*((PROTOCOL_ID*)&buffer[offset]));
-    offset += 2;
+    /// get version
+    offset += sizeof(m_length);
+    m_version = ntohs(*((VERSION_TYPE*)&buffer[offset]));
+    /// get protocolID
+    offset += sizeof(m_version);
+    m_protocolID = ntohl(*((PROTOCOL_ID*)&buffer[offset]));
+    /// get pacektType
+    offset += sizeof(m_protocolID);
     m_packetType = ntohs(*((PACKET_TYPE*)&buffer[offset]));
-    offset += 2;
+    /// get sesq
+    offset += sizeof(m_packetType);
     m_seq = ntohl(*((uint32_t*)&buffer[offset]));
-    ///< TODO: assign to std::move
-    m_buffer->assign(&buffer[HEADER_LENGTH], &buffer[HEADER_LENGTH] + m_length - HEADER_LENGTH);
 
+    /// the data has been compressed
+    if (g_BCOSConfig.compressEnabled() &&
+        ((m_version & dev::eth::CompressFlag) == dev::eth::CompressFlag))
+    {
+        /// uncompress data
+        SnappyCompress::uncompress(
+            bytesConstRef((const byte*)(&buffer[HEADER_LENGTH]), m_length - HEADER_LENGTH),
+            *m_buffer);
+    }
+    else
+    {
+        ///< TODO: assign to std::move
+        m_buffer->assign(&buffer[HEADER_LENGTH], &buffer[HEADER_LENGTH] + m_length - HEADER_LENGTH);
+    }
     return m_length;
-}
-
-void P2PMessage::encodeAMOPBuffer(std::string const& topic)
-{
-    ///< check protocolID is AMOP message or not
-    if (dev::eth::ProtocolID::Topic != abs(m_protocolID))
-    {
-        return;
-    }
-
-    ///< new buffer format:topic lenght + topic data + ori buffer data
-    m_buffer->insert(m_buffer->begin(), topic.begin(), topic.end());
-    uint32_t topicLen = htonl(topic.size());
-    m_buffer->insert(m_buffer->begin(), (byte*)&topicLen, (byte*)&topicLen + sizeof(topicLen));
-}
-
-ssize_t P2PMessage::decodeAMOPBuffer(std::shared_ptr<bytes> buffer, std::string& topic)
-{
-    ///< check protocolID is AMOP message or not
-    if (dev::eth::ProtocolID::Topic != abs(m_protocolID))
-    {
-        return dev::network::PACKET_ERROR;
-    }
-
-    uint32_t topicLen = ntohl(*((uint32_t*)m_buffer->data()));
-    P2PMSG_LOG(TRACE) << LOG_DESC("Message::decodeAMOPBuffer topic") << LOG_KV("len", topicLen)
-                      << LOG_KV("buffer size", m_buffer->size());
-    if (topicLen + 4 > m_buffer->size())
-    {
-        return dev::network::PACKET_ERROR;
-    }
-    topic = std::string((char*)(m_buffer->data()) + 4, topicLen);
-    buffer->insert(buffer->end(), m_buffer->begin() + 4 + topicLen, m_buffer->end());
-
-    return buffer->size();
 }
